Fix coverage print problem in multi-project
In multi-project, header (Coverage summary:) should be printed
for the first subproject build, but when some project is not included
in the task graph, header is not shown or coverage summary line
is incorrectly indented.

  $ ./gradlew project2:jacocoTestReport
  -> Lack of header
    Total time: 5.525 secs
    project2-with-long-name: 44.4%

  $ ./gradlew project1:jacocoTestReport
  -> incorrect indent
    Total time: 6.295 secs

    Coverage summary:
    project1:                72.2%

With this change, only the projects that are included in the task graph
will be checked, and fix the issue.